FIX(catalog-page): Change the current page when the user applies filters
Jira: [[FE]Change the current page when the user applies filters](https://cozyhometeam.atlassian.net/browse/CH-486)
